Separate policy schema files
